Create README.md
A Blood Donation System is a database system that helps manage the collection, storage, tracking, and distribution of blood donations. This system facilitates the interaction between donors, donations, blood _request, blood _inventory, transfusions and other entities involved in the blood donation process. It aims to ensure that blood supplies are adequate, track blood types, and improve efficiency in matching donors with patients in need.